+++ c++direct2dapp/c++direct2dapp/Platform.cpp
@@ -2,7 +2,7 @@
 
 Platform::Platform() {}
 
-Platform::Platform(Point _position, int _width, int _height, float _yVel, float _xVel, Graphics * gfxi) :
+Platform::Platform(Point _position, float _width, float _height, float _yVel, float _xVel, Graphics * gfxi) :
 	Rect(_position, _width, _height, _yVel, _xVel, true, platform, gfxi)
 {
