+++ c++direct2dapp/c++direct2dapp/Platform.h
@@ -5,6 +5,6 @@ class Platform : public Rect {
 
 public:
 	Platform();
-	Platform(Point _position, int _width, int _height, float _yVel, float _xVel, Graphics * gfxi);
+	Platform(Point _position, float _width, float _height, float _yVel, float _xVel, Graphics * gfxi);
 	~Platform();
 };
